Finished field and sensorarray first pass, put everything into a mono repo
