+++ src/pycave/__init__.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
 '''
--------------------------------------------------------------------------------
+================================================================================
 pycave
 
 authors: thescepticalrabbit
--------------------------------------------------------------------------------
+================================================================================
 '''
 
+# IMPORT THE MONO REPO!!!
+from pycave.pycave import *
 
-
-__all__ = []
+__all__ = ["pycave"]
