--- src/pycave/measurement.py
@@ -1,7 +1,0 @@
-'''
--------------------------------------------------------------------------------
-pycave.measurement
-
-authors: thescepticalrabbit
--------------------------------------------------------------------------------
-'''
