--- src/pycave/sensor.py
@@ -1,7 +1,0 @@
-'''
--------------------------------------------------------------------------------
-pycave.sensor
-
-authors: thescepticalrabbit
--------------------------------------------------------------------------------
-'''
