--- src/pycave/sensorarray.py
@@ -1,7 +1,0 @@
-'''
--------------------------------------------------------------------------------
-pycave.sensorarray
-
-authors: thescepticalrabbit
--------------------------------------------------------------------------------
-'''
